Add Earn Points flow and use double amounts
Implements the Earn Points feature and related refactors: adds an earn_points module (repository, models: ResolvedGuest, PointsPreview, EarnTransactionRequest), a comprehensive Notifier controller for scanning/preview/submission, and the Bill Details UI/page and supporting widgets. Integrates a new /staff/earn-points route and updates the staff dashboard to push the flow. Changes earning rule amount types from int to double across controllers, pages and widgets to support fractional amounts. Also updates API baseUrl and adds mobile_scanner entries to IDE/project metadata.

+++ beloyal_app/lib/features/business_loyalty/data/repositories/earning_rule_repository.dart
@@ -8,7 +8,7 @@ class EarningRuleRepository {
 
   Future<void> patchEarningSettings({
     required int businessId,
-    required int amountPer,
+    required double amountPer,
     required int pointsPer,
   }) async {
     await _dio.patch(
